backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
idna==3.4
inflection==0.5.1
ipykernel==6.26.0
ipython==8.17.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
phonenumbers==8.13.24
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
slugify==0.0.1
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
